Jamming efficiency
I made these changes to measure how different jamming strategies
block payments in the LN.

The simulator itself is pretty massive and complex. I suggest to just
looking at the diff of this commit first, and then whatever is called by it,
rather than inspecting everything.

This code is heavily research-oriented, and likely to contain bugs. However,
the results look sane and I expect them to be close to reality (except certain
simplifications discussed in the research).